# spring/api-docs.yaml
openapi: 3.0.1
info:
  title: Конспект по языку Java
  description: Конспект по языку Java на русском языке
  contact:
    name: Токовенко Виктор
    url: https://github.com/Lorgar-Avrelian?tab=repositories
    email: victor-14-244@mail.ru
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /manage:
    get:
      tags:
        - Приём/выдача
        - 4 Приём/выдача книг
      summary: Принять
      description: Принять книгу от читателя
      operationId: takeBook
      parameters:
        - name: book
          in: query
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: reader
          in: query
          description: ID читателя в имеющемся списке читателей
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reader"
        "400":
          description: Bad Request
      security:
        - basicAuth: []
    post:
      tags:
        - Приём/выдача
        - 4 Приём/выдача книг
      summary: Выдать
      description: Выдать книгу читателю
      operationId: giveBook
      parameters:
        - name: bookId
          in: query
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: readerId
          in: query
          description: ID читателя в имеющемся списке читателей
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reader"
        "400":
          description: Bad Request
      security:
        - basicAuth: []
...
  /expenses:
    get:
      tags:
        - 6 Затраты
        - Затраты
      summary: Все затраты
      description: Посмотреть все затраты
      operationId: getExpenses
      parameters:
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: size
          in: query
          description: Размер страницы
          required: true
          schema:
            type: integer
            format: int64
          example: 4
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      security:
        - basicAuth: []
    post:
      tags:
        - 6 Затраты
        - Затраты
      summary: Внести
      description: Внести затрату
      operationId: createExpense
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Expense"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Bad Request
      security:
        - basicAuth: []
...
  /expenses/{id}:
    get:
      tags:
        - 6 Затраты
        - Затраты
      summary: Посмотреть
      description: Посмотреть затрату по ID
      operationId: getExpenseById
      parameters:
        - name: id
          in: path
          description: ID записи о затратах
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not Found
      security:
        - basicAuth: []
  /expenses/categories:
    get:
      tags:
        - 6 Затраты
        - Затраты
      summary: Затраты по категориям
      description: Посмотреть все затраты по категориям
      operationId: getExpensesByCategories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpensesByCategory"
        "404":
          description: Not Found
      security:
        - basicAuth: []
...
  /counter:
    get:
      tags:
        - 7 Счётчик
      summary: Получить
      description: Получить значение счётчика
      operationId: getCounter
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "400":
          description: Bad Request
      security:
        - basicAuth: []
  /counter/change:
    get:
      tags:
        - 7 Счётчик
      summary: Изменить - вариант 1
      description: Изменить значение счётчика - вариант 1
      operationId: changeCounter1
      parameters:
        - name: counter
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "400":
          description: Bad Request
      security:
        - basicAuth: []
  /counter/change/{counter}:
    get:
      tags:
        - 7 Счётчик
      summary: Изменить - вариант 2
      description: Изменить значение счётчика - вариант 2
      operationId: changeCounter2
      parameters:
        - name: counter
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "400":
          description: Bad Request
      security:
        - basicAuth: []
...
components:
  schemas:
...
    Reader:
      title: Читатель
      maximum: 9223372036854775807
      minimum: 1
      required:
        - id
        - name
        - personalNumber
        - surname
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID читателя
          format: int64
          default: 1
        name:
          title: Имя
          maxLength: 12
          minLength: 2
          type: string
          description: Имя читателя
          default: Иван
        secondName:
          title: Отчество
          maxLength: 16
          minLength: 2
          type: string
          description: Отчество читателя
          default: Иванович
        surname:
          title: Фамилия
          maxLength: 30
          minLength: 2
          type: string
          description: Фамилия читателя
          default: Иванов
        personalNumber:
          title: Номер билета
          maximum: 2147483648
          minimum: 1
          type: integer
          description: Номер читательского билета
          format: int32
          default: 1
      description: Сущность читателя
      default: null
...
    Expense:
      title: Расход
      required:
        - amount
        - category
        - date
        - id
        - title
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID траты
          format: int32
          default: 1
        title:
          title: Название
          maxLength: 30
          minLength: 3
          type: string
          description: Название траты
          default: Поездка на транспорте
        date:
          title: Дата
          type: string
          description: Дата траты
          format: date
        category:
          title: Категория
          maxLength: 30
          minLength: 3
          type: string
          description: Категория затрат
          default: Транспорт
        amount:
          title: Количество
          minimum: 0
          type: number
          description: Количество затрат
          format: float
          default: 1000
      description: Сущность затрат
...
    ExpensesByCategory:
      title: Расходы по категориям
      type: object
      properties:
        amount:
          title: Количество
          minimum: 0
          type: number
          description: Количество затрат
          format: float
          default: 1000
        category:
          title: Категория
          type: string
          description: Название категории
          default: Транспорт
      description: Сущность расходов по категориям
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic